+++ src/queryEditorProvider.ts
@@ -8,6 +8,7 @@ export class QueryEditorProvider {
 		if (!this.context) {
 		}
 	}
+
 	async commandGenerateLql(uri: vscode.Uri) {
 		if (!uri) {
 			return Promise.resolve();
@@ -15,7 +16,7 @@ export class QueryEditorProvider {
 		if (uri.scheme !== 'file') {
 			return vscode.window.showErrorMessage('Only local files converted to LQL.');
 		}
-		const account = StateManager.account;
+		const account = StateManager.getActiveAccount();
 		if (!account) {
 			return vscode.window.showErrorMessage('Connect to an account before converting a file to LQL.');
 		}
@@ -53,29 +54,40 @@ export class QueryEditorProvider {
 		});
 	}
 
-	async commandUploadQuery(uri: vscode.Uri) {
-		if (!uri) {
-			return Promise.resolve();
-		}
-		if (uri.scheme !== 'file') {
-			return vscode.window.showErrorMessage('Only local files can be uploaded to Lytics.');
-		}
-		const account = StateManager.account;
+	async commandUploadQuery(uri: vscode.Uri): Promise<boolean> {
+		// await vscode.window.withProgress({
+		// 	location: vscode.ProgressLocation.Notification,
+		// 	title: `Uploading query.`,
+		// 	cancellable: true
+		// }, async (progress, token) => {
+		// });
+		const account = StateManager.getActiveAccount();
 		if (!account) {
-			return vscode.window.showErrorMessage('Connect to an account before uploading a query to Lytics.');
+			vscode.window.showErrorMessage('Connect to an account before uploading a query to Lytics.');
+			return Promise.resolve(false);
 		}
-		const editor = vscode.window.activeTextEditor;
-		if (editor && editor.document) {
-			if (editor.document.uri.toString() !== uri.toString()) {
-				//this should not happen
-				return vscode.window.showErrorMessage('Editor and query documents do not match.');
+		if (!uri) {
+			const selectedFiles = await vscode.window.showOpenDialog({
+				canSelectFiles: true,
+				canSelectFolders: false,
+				canSelectMany: false,
+			});
+			if (!selectedFiles || selectedFiles.length !== 1) {
+				return Promise.resolve(false);
 			}
-			if (editor.document.isDirty) {
-				return vscode.window.showErrorMessage('Save the file before uploading a query to Lytics.');
-			}
+			uri = selectedFiles[0];
 		}
-		//get the query alias from the file
-		fs.readFile(uri.fsPath, 'utf-8', async (err, data) => {
+		const doc = vscode.workspace.textDocuments.find(doc => doc.uri.fsPath === uri.fsPath);
+		if (uri.scheme !== 'file') {
+			vscode.window.showErrorMessage('Only local files can be uploaded to Lytics.');
+			return Promise.resolve(false);
+		}
+		if (doc.isDirty) {
+			vscode.window.showErrorMessage('Save the file before uploading a query to Lytics.');
+			return Promise.resolve(false);
+		}
+		var wasUploaded = false;
+		await fs.readFile(uri.fsPath, 'utf-8', async (err, data) => {
 			const regex = /\bALIAS\s*([a-zA-Z_\-]+)/;
 			const matches = data.match(regex);
 			let alias;
@@ -84,8 +96,9 @@ export class QueryEditorProvider {
 			}
 			if (!alias || alias.length === 0) {
 				await vscode.window.showErrorMessage('Unable to upload query: no query alias was found in the LQL.');
-				return Promise.resolve();
+				return;
 			}
+			//
 			//determine if the query already exists
 			var currentQuery;
 			const client = lytics.getClient(account.apikey!);
@@ -95,23 +108,25 @@ export class QueryEditorProvider {
 			catch (err) {
 				if (!err || !err.response || !err.response.status || err.response.status !== 404) {
 					vscode.window.showErrorMessage(`Unable to determine if the query already exists.`);
-					return Promise.resolve();
+					return;
 				}
 			}
+			//
 			//prompt the user to add new or overwrite existing query
 			const yesMessage = currentQuery ? 'Overwrite existing query on Lytics' : 'Add new query to Lytics';
 			const noMessage = 'Cancel';
-			const answer = await vscode.window.showQuickPick([yesMessage, noMessage], { canPickMany: false });
+			const answer = await vscode.window.showQuickPick([noMessage, yesMessage], { canPickMany: false });
 			if (answer !== yesMessage) {
-				return Promise.resolve();
+				return;
 			}
 			const response = await client.upsertQuery(data);
 			if (response.length < 1) {
 				vscode.window.showErrorMessage(`Unable to upload query ${alias} to Lytics}.`);
-				return Promise.resolve();
+				return;
 			}
 			vscode.window.showInformationMessage(`Query ${alias} was uploaded to Lytics.`);
-			return Promise.resolve();
+			wasUploaded = true;
 		});
+		return Promise.resolve(wasUploaded);
 	}
 }