+++ src/streamExplorerProvider.ts
@@ -14,8 +14,12 @@ export class StreamExplorerProvider extends LyticsExplorerProvider<DataStream | 
 
 	private mapOfFieldToStream: Map<DataStreamField, DataStream> = new Map<DataStreamField, DataStream>();
 
+	/**
+	 * Inherited from TreeDataProvider.
+	 * @param element 
+	 */
 	async getChildren(element?: DataStream): Promise<any[]> {
-		const account = StateManager.account;
+		const account = StateManager.getActiveAccount();
 		if (!account) {
 			return Promise.resolve([]);
 		}
@@ -37,32 +41,44 @@ export class StreamExplorerProvider extends LyticsExplorerProvider<DataStream | 
 		return Promise.resolve([]);
 	}
 
-	async getParent(element?: any): Promise<any> {
-		if (!element) {
-			return Promise.resolve(undefined);
-		}
-		if (element.name) {
-			return Promise.resolve(this.mapOfFieldToStream.get(element));
-		}
-		return Promise.resolve(undefined);
-	}
-
+	/**
+	 * Inherited from TreeDataProvider.
+	 * @param element 
+	 */
 	getTreeItem(element: any): vscode.TreeItem {
 		if (element.stream) {
 			return new DataStreamTreeItem(element);
 		}
 		if (element.name) {
 			let item = new DataStreamFieldTreeItem(element);
-			item.iconPath = this.getIcon(element);
+			item.iconPath = this.getStreamFieldIcon(element);
 			return item;
 		}
 		throw new Error(`The specified element is not supported by the stream explorer provider.`);
 	}
 
-	private getIcon(element: DataStreamField): any {
+	/**
+	 * Gets the parent for the node.
+	 * @param node 
+	 */
+	async getParent(node?: any): Promise<any> {
+		if (!node) {
+			return Promise.resolve(undefined);
+		}
+		if (node.name) {
+			return Promise.resolve(this.mapOfFieldToStream.get(node));
+		}
+		return Promise.resolve(undefined);
+	}
+
+	/**
+	 * Gets the icon for a data stream field node.
+	 * @param node 
+	 */
+	private getStreamFieldIcon(node: any): any {
 		var icon: (string | undefined) = undefined;
-		if (element instanceof DataStreamField) {
-			icon = `${element.type}.svg`;
+		if (node.type !== undefined) {
+			icon = `${node.type}.svg`;
 		}
 		if (!icon) {
 			return undefined;
@@ -73,6 +89,36 @@ export class StreamExplorerProvider extends LyticsExplorerProvider<DataStream | 
 		};
 	}
 
+	/**
+	 * Displays a quick pick from which the user can select a data stream.
+	 * @param account 
+	 * @param message 
+	 * @returns The selected stream, or undefined if none was selected.
+	 */
+	private async promptForStream(account: LyticsAccount, message?: string): Promise<DataStream | undefined> {
+		if (!message) {
+			message = `Select a data stream.`;
+		}
+		const streams = await this.getStreams(account);
+		const values = streams.map(q => q.stream);
+		let value = await vscode.window.showQuickPick(values, {
+			canPickMany: false,
+			placeHolder: message
+		});
+		if (!value) {
+			return Promise.resolve(undefined);
+		}
+		const stream = streams.find(s => s.stream === value);
+		return Promise.resolve(stream);
+	}
+
+	/**
+	 * Wrapper around the API call to get Lytics data streams. 
+	 * This function provides user feedback while data is 
+	 * read from Lytics.
+	 * @param account 
+	 * @returns Array of data streams.
+	 */
 	async getStreams(account: LyticsAccount): Promise<DataStream[]> {
 		let streams = await vscode.window.withProgress({
 			location: vscode.ProgressLocation.Notification,
@@ -129,11 +175,45 @@ export class StreamExplorerProvider extends LyticsExplorerProvider<DataStream | 
 		return Promise.resolve(fields);
 	}
 
-	async commandShowQueryInfo(stream: DataStream) {
+	async commandShowStreamInfo(stream: DataStream): Promise<boolean> {
 		try {
-			const account = StateManager.account;
+			const account = StateManager.getActiveAccount();
 			if (!account) {
 				throw new Error('No account is connected.');
+			}
+			if (!stream) {
+				stream = await this.promptForStream(account, `Select the data stream you want to show.`);
+			}
+			if (!stream) {
+				return Promise.resolve(false);
+			}
+			await vscode.window.withProgress({
+				location: vscode.ProgressLocation.Notification,
+				title: `Loading stream info: ${stream.stream}`,
+				cancellable: true
+			}, async (progress, token) => {
+				const uri = vscode.Uri.parse(`lytics://${account.aid}/streams/${stream.stream}.json`);
+				await this.displayAsReadOnly(uri);
+			});
+			return Promise.resolve(true);
+		}
+		catch (err) {
+			vscode.window.showErrorMessage(`Loading data stream info failed: ${err.message}`);
+			return Promise.resolve(false);
+		}
+	}
+
+	async commandShowQueryInfo(stream: DataStream) {
+		try {
+			const account = StateManager.getActiveAccount();
+			if (!account) {
+				throw new Error('No account is connected.');
+			}
+			if (!stream) {
+				stream = await this.promptForStream(account, `Select the data stream whose query info you want to show.`);
+			}
+			if (!stream) {
+				return Promise.resolve(false);
 			}
 			await vscode.window.withProgress({
 				location: vscode.ProgressLocation.Notification,
@@ -152,10 +232,11 @@ export class StreamExplorerProvider extends LyticsExplorerProvider<DataStream | 
 
 	async commandShowField(field: DataStreamField) {
 		try {
-			const account = StateManager.account;
+			const account = StateManager.getActiveAccount();
 			if (!account) {
 				throw new Error('No account is connected.');
 			}
+			//TODO: if field is udefined...
 			const stream = await this.getParent(field) as DataStream;
 			if (!stream) {
 				throw new Error(`No parent was found for field ${field.name}`);